Reorganize the organization of expcetions
