+++ plugins/edu.uoc.som.wapiml.ui/src/edu/uoc/som/wapiml/ui/handlers/ConvertToYaml.java
@@ -20,7 +20,7 @@ import org.eclipse.jface.viewers.ISelection;
 import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.ui.handlers.HandlerUtil;
 
-import edu.uoc.som.openapi2.io.exception.OpenAPIValidationException;
+import edu.uoc.som.openapi2.io.exceptions.OpenAPIValidationException;
 import edu.uoc.som.openapi2.io.model.SerializationFormat;
 import edu.uoc.som.wapiml.ui.WAPImlUIPlugin;
 import edu.uoc.som.wapiml.utils.IOUtils;
@@ -66,8 +66,6 @@ public class ConvertToYaml extends AbstractHandler {
 						return new Status(IStatus.ERROR, WAPImlUIPlugin.PLUGIN_ID, e.getLocalizedMessage(),
 								e.getCause());
 
-					} catch (OpenAPIValidationException e) {
-						return new Status(IStatus.ERROR, WAPImlUIPlugin.PLUGIN_ID, e.getLocalizedMessage());
 					} finally {
 						monitor.done();
 					}
